Update beer model to fix compatibility with API

+++ lib/domain/beers/value_objects/beer_ebc.dart
@@ -1,5 +1,5 @@
 import 'package:beer_app/domain/core/value_object.dart';
 
-class BeerEbc extends ValueObject<int> {
+class BeerEbc extends ValueObject<double> {
   const BeerEbc(super.value);
 }
