+++ lib/domain/beers/value_objects/beer_srm.dart
@@ -1,5 +1,5 @@
 import 'package:beer_app/domain/core/value_object.dart';
 
-class BeerSrm extends ValueObject<int> {
+class BeerSrm extends ValueObject<double> {
   const BeerSrm(super.value);
 }
