+++ lib/domain/beers/value_objects/beer_target_fg.dart
@@ -1,5 +1,5 @@
 import 'package:beer_app/domain/core/value_object.dart';
 
-class BeerTargetFg extends ValueObject<int> {
+class BeerTargetFg extends ValueObject<double> {
   const BeerTargetFg(super.value);
 }
