+++ lib/domain/beers/value_objects/beer_target_og.dart
@@ -1,5 +1,5 @@
 import 'package:beer_app/domain/core/value_object.dart';
 
-class BeerTargetOg extends ValueObject<int> {
+class BeerTargetOg extends ValueObject<double> {
   const BeerTargetOg(super.value);
 }
